Fixes for qt in Python 3.10 from electron-cash (#10)
* Fixup for Python 3.10

This commit contains some fixups for Python 3.10.  Apparently some of
the PyQt overloads disappeared that support int <-> float
interchangeably as args to C++ functions, which caused the app to not
work at all on Python 3.10.

Note: there are still extant bugs with using Python 3.10, but this
commit at least allows the app to (mostly) run ok.

* More fixups for Python 3.10

For some reason now, PyQt5 doesn't auto-coerce args from int -> float
and float -> int, so everything that is passing the wrong type from
Python to PyQt is broken now.  Luckily this doesn't happen in too many
places, but it does happen.

This commit fixes more of these errors leading to app crash.

* Python 3.10 fixups

Fixed a bunch of places where potentially non-bool or non-int params are
being passed to functions expecting precisely bools and/or ints.

Also fixed a typo in main_window.py.

+++ oregano_gui/qt/qrreader/video_overlay.py
@@ -66,7 +66,7 @@ class QrReaderVideoOverlay(QWidget):
         self.bg_rect_pen = QPen()
         self.bg_rect_pen.setColor(Qt.black)
         self.bg_rect_pen.setStyle(Qt.DotLine)
-        self.bg_rect_fill = QColor(255, 255, 255, 255 * self.BG_RECT_OPACITY)
+        self.bg_rect_fill = QColor(255, 255, 255, int(255 * self.BG_RECT_OPACITY))
 
         self.qr_finder = QSvgRenderer(":icons/qr_finder.svg")
 
@@ -99,12 +99,12 @@ class QrReaderVideoOverlay(QWidget):
         # Compute the transform to flip the coordinate system on the x axis
         transform_flip = QTransform()
         if self.flip_x:
-            transform_flip = transform_flip.translate(self.resolution.width(), 0.0)
+            transform_flip = transform_flip.translate(float(self.resolution.width()), 0.0)
             transform_flip = transform_flip.scale(-1.0, 1.0)
 
         # Small helper for tuple to QPoint
         def toqp(point):
-            return QPoint(point[0], point[1])
+            return QPoint(int(point[0]), int(point[1]))
 
         # Starting from here we care about AA
         painter.setRenderHint(QPainter.Antialiasing)
